requirements.txt
litellm
google-generativeai
cohere
faiss-cpu